add basic http test
In order to prevent ring/jetty from throwing a 500, `pr-str` is called over the
result of calling `node/list-tasks`. Seemingly this is necessary to prevent
errors when serving the request. However there might be a better way of
handling this and simply serving up bytes directly?

This patch also adds a 404 response, when no route matches.